feat: Drop deprecated hook script support (issue #62)
